backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

sqlalchemy
aiosqlite            # para SQLite

langchain
langchain-community  # integra Ollama en LangChain
